Suggest database to use pl_namespace index for link counting
Our prod mysql servers are generating a poor query plan for this
query that only uses the index for the namespace lookup and not the
title. I can't particularly explain why (the index mysql chooses and
the index provided here have the same first two fields, which are
the two we are querying), but when using this index mysql chooses
to use the index for the pl_title field as well which runs the
query orders of magnitude faster.

Bug: T143932
Change-Id: I0b648475aee2e3d64eeaad02b3c25f9c300c9602
(cherry picked from commit f75c3f147691173f624769a4cdbac575c5f17f3c)