Record server hostname in CirrusSearchRequestSet
When tracking down timing regressions it would be nice to know if those
regressions are localized to a specific set of servers, or if it happens
across the board. Record the hostname in CSRQ to help with this.

Change-Id: I2ea70bb596e1d0ff4b050fafe709798bc4a8ddf9
(cherry picked from commit 8c905291011f71d51f7ab795e1f453e5a8d6632f)