Remove the standard token filter
It does nothing, elasticsearch docs states that it's just a placeholder
in case future filtering options needs to be added there.
I think it's safer to remove it and re-add it explicitly if these
"future" filtering options are useful to us.
It helps also if we want to plug another tokenizer to replace this
"standard" analysis chain.

Change-Id: I4d64c681119b02634b5312429b30bb5010d8f111