Add wgContentNamespaces to the list of vars loaded by SearchConfig
This is our best bet to explain the errors that started to spam the
logs when we activated BM25 on the top10 wikipedias.
It was the first time that a rescore profile was used with the value
'content' as supported_namespaces, this tells the Rescorebuilder to
inspect wgContentNamespaces.
Unfortunately when running a secondTry query issued from a TextCat
language detection we load an external wiki config via SearchConfig.
This class uses a prefix filter for cirrus vars and a whitelist
approach for non cirrus one.

This patch adds 'wgContentNamespaces' to this whitelist.

Bug: T148840
Change-Id: If7452e0d12dd0cb3d3fed36106708e5a0f95c41d
(cherry picked from commit 47bd6747a1537ab85d0443396c06e92ab4a674c8)